requirements.txt
fastapi
motor
openai
pydantic
pymongo
PyPDF2
python-dotenv
python_docx
pywin32; sys_platform == 'win32'
uvicorn
python-multipart
starlette

